Change the way number of columns are calculated
It's a lot simpler to get the item width and calculate number of
columns based on the window width. The previous implementation was
meant to be an optimization, but was more error prone than this
one.

This fixes a few severe bugs with window resize calculations and
dynamic loading of data